requirements.txt
boto3==1.42.74
dotenv==0.9.9
pyspark==3.4.2
py4j==0.10.9.7
kafka-python==2.3.0
pymongo==4.16.0
clickhouse-connect==0.8.0
airflow-providers-clickhouse==0.0.1
clickhouse-driver==0.2.9
airflow-clickhouse-plugin==1.6.0